+++ devel/undocumented_fields/undocumented_fields.txt
@@ -393,7 +393,7 @@
     - quantity_type
     - tensor_type
 
-  Module: ansys.fluent.core.utils.file_transfer_service
+  Module: ansys.fluent.core.file_transfer_service
   Class: ContainerFileTransferStrategy
     - client
     - container
