+++ src/ansys/fluent/core/file_transfer_service.py
@@ -90,7 +90,7 @@ class StandaloneFileTransferStrategy(FileTransferStrategy):
     --------
     >>> import ansys.fluent.core as pyfluent
     >>> from ansys.fluent.core import examples
-    >>> from ansys.fluent.core.utils.file_transfer_service import StandaloneFileTransferStrategy
+    >>> from ansys.fluent.core.file_transfer_service import StandaloneFileTransferStrategy
     >>> mesh_file_name = examples.download_file("mixing_elbow.msh.h5", "pyfluent/mixing_elbow")
     >>> meshing_session = pyfluent.launch_fluent(mode=pyfluent.FluentMode.MESHING, file_transfer_service=StandaloneFileTransferStrategy())
     >>> meshing_session.upload(file_name=mesh_file_name, remote_file_name="elbow.msh.h5")
@@ -148,7 +148,7 @@ class StandaloneFileTransferStrategy(FileTransferStrategy):
         --------
         >>> import ansys.fluent.core as pyfluent
         >>> from ansys.fluent.core import examples
-        >>> from ansys.fluent.core.utils.file_transfer_service import StandaloneFileTransferStrategy
+        >>> from ansys.fluent.core.file_transfer_service import StandaloneFileTransferStrategy
         >>> mesh_file_name = examples.download_file("mixing_elbow.msh.h5", "pyfluent/mixing_elbow")
         >>> meshing_session = pyfluent.launch_fluent(mode=pyfluent.FluentMode.MESHING, file_transfer_service=StandaloneFileTransferStrategy())
         >>> meshing_session.upload(file_name=mesh_file_name, remote_file_name="elbow.msh.h5")
@@ -183,7 +183,7 @@ class StandaloneFileTransferStrategy(FileTransferStrategy):
         --------
         >>> import ansys.fluent.core as pyfluent
         >>> from ansys.fluent.core import examples
-        >>> from ansys.fluent.core.utils.file_transfer_service import StandaloneFileTransferStrategy
+        >>> from ansys.fluent.core.file_transfer_service import StandaloneFileTransferStrategy
         >>> mesh_file_name = examples.download_file("mixing_elbow.msh.h5", "pyfluent/mixing_elbow")
         >>> meshing_session = pyfluent.launch_fluent(mode=pyfluent.FluentMode.MESHING, file_transfer_service=StandaloneFileTransferStrategy())
         >>> meshing_session.meshing.File.WriteMesh(FileName="write_elbow.msh.h5")
@@ -224,7 +224,7 @@ class ContainerFileTransferStrategy(FileTransferStrategy):
     --------
     >>> import ansys.fluent.core as pyfluent
     >>> from ansys.fluent.core import examples
-    >>> from ansys.fluent.core.utils.file_transfer_service import ContainerFileTransferStrategy
+    >>> from ansys.fluent.core.file_transfer_service import ContainerFileTransferStrategy
     >>> case_file_name = examples.download_file("mixing_elbow.cas.h5", "pyfluent/mixing_elbow")
     >>> solver_session = pyfluent.launch_fluent(file_transfer_service=ContainerFileTransferStrategy())
     >>> solver_session.upload(file_name=case_file_name, remote_file_name="elbow.cas.h5")
@@ -358,7 +358,7 @@ class ContainerFileTransferStrategy(FileTransferStrategy):
         --------
         >>> import ansys.fluent.core as pyfluent
         >>> from ansys.fluent.core import examples
-        >>> from ansys.fluent.core.utils.file_transfer_service import ContainerFileTransferStrategy
+        >>> from ansys.fluent.core.file_transfer_service import ContainerFileTransferStrategy
         >>> case_file_name = examples.download_file("mixing_elbow.cas.h5", "pyfluent/mixing_elbow")
         >>> solver_session = pyfluent.launch_fluent(file_transfer_service=ContainerFileTransferStrategy())
         >>> solver_session.upload(file_name=case_file_name, remote_file_name="elbow.cas.h5")
@@ -399,7 +399,7 @@ class ContainerFileTransferStrategy(FileTransferStrategy):
         --------
         >>> import ansys.fluent.core as pyfluent
         >>> from ansys.fluent.core import examples
-        >>> from ansys.fluent.core.utils.file_transfer_service import ContainerFileTransferStrategy
+        >>> from ansys.fluent.core.file_transfer_service import ContainerFileTransferStrategy
         >>> case_file_name = examples.download_file("mixing_elbow.cas.h5", "pyfluent/mixing_elbow")
         >>> solver_session = pyfluent.launch_fluent(file_transfer_service=ContainerFileTransferStrategy())
         >>> solver_session.settings.file.write_case(file_name="write_elbow.cas.h5")
