+++ src/ansys/fluent/core/launcher/pim_launcher.py
@@ -45,6 +45,7 @@ from typing import TYPE_CHECKING, Any, TypedDict
 from typing_extensions import Unpack
 
 from ansys.fluent.core._types import LauncherArgsBase
+from ansys.fluent.core.file_transfer_service import PimFileTransferService
 from ansys.fluent.core.fluent_connection import FluentConnection, _get_max_c_int_limit
 from ansys.fluent.core.launcher.launch_options import (
     Dimension,
@@ -52,7 +53,6 @@ from ansys.fluent.core.launcher.launch_options import (
     _get_argvals_and_session,
 )
 from ansys.fluent.core.session.session import _parse_server_info_file
-from ansys.fluent.core.utils.file_transfer_service import PimFileTransferService
 from ansys.fluent.core.utils.fluent_version import FluentVersion
 import ansys.platform.instancemanagement as pypim
