+++ tests/test_launcher_remote.py
@@ -41,6 +41,7 @@ from test_session import MockHealthServicer, MockSchemeEvalServicer
 from ansys.api.fluent.v0 import scheme_eval_pb2_grpc
 import ansys.fluent.core as pyfluent
 from ansys.fluent.core import config, examples
+from ansys.fluent.core.file_transfer_service import PimFileTransferService
 from ansys.fluent.core.fluent_connection import (
     FluentConnection,
     UnsupportedRemoteFluentInstance,
@@ -49,7 +50,6 @@ from ansys.fluent.core.launcher import launcher
 from ansys.fluent.core.session.pure_meshing import PureMeshing
 from ansys.fluent.core.session.session import BaseSession
 from ansys.fluent.core.session.solver import Solver
-from ansys.fluent.core.utils.file_transfer_service import PimFileTransferService
 import ansys.fluent.core.utils.fluent_version as docker_image_version
 from ansys.fluent.core.utils.fluent_version import FluentVersion
 from ansys.fluent.core.utils.networking import get_free_port
